+++ pathfinder2e_stats/armory/starfinder/dart.py
@@ -1,5 +1,0 @@
-from pathfinder2e_stats.armory._common import _weapon
-
-__all__ = ("force_needle",)
-
-force_needle = _weapon("force needle", "piercing", 4)