+++ pathfinder2e_stats/armory/starfinder/flame.py
@@ -1,5 +1,0 @@
-from pathfinder2e_stats.armory._common import _weapon
-
-__all__ = ("skyfire_sword",)
-
-skyfire_sword = _weapon("skyfire sword", "fire", 8, two_hands=10)