+++ pathfinder2e_stats/armory/starfinder/hammer.py
@@ -1,5 +1,0 @@
-from pathfinder2e_stats.armory._common import _weapon
-
-__all__ = ("hammer",)
-
-hammer = _weapon("hammer", "bludgeoning", 8)