+++ pathfinder2e_stats/armory/starfinder/shock.py
@@ -1,6 +1,0 @@
-from pathfinder2e_stats.armory._common import _weapon
-
-__all__ = ("polyglove", "shock_pad")
-
-polyglove = _weapon("polyglove", "electricity", 6)
-shock_pad = _weapon("shock pad", "electricity", 4)